third_party/WebKit/LayoutTests/fast/backgrounds/background-clip-per-layer.html

<!DOCTYPE html>
<html>
<script src="../../resources/js-test.js"></script>
<div id="test"> </div>
<script>
description("Tests that background shorthand property parsing will deal correctly with clip values in multiple layers.");

var e = document.getElementById('test');

e.style.background = "url('a') content-box padding-box, url('b') border-box";
shouldBeEqualToString("e.style.backgroundOrigin", 'content-box, border-box');
shouldBeEqualToString("e.style.backgroundClip", 'padding-box, border-box');

</script>
</body>
</html>
