third_party/WebKit/LayoutTests/fast/backgrounds/background-clip-per-layer-expected.txt

Tests that background shorthand property parsing will deal correctly with clip values in multiple layers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS e.style.backgroundOrigin is "content-box, border-box"
PASS e.style.backgroundClip is "padding-box, border-box"
PASS successfullyParsed is true

TEST COMPLETE
